LayoutTests/sputnik/Conformance/08_Types/8.4_The_String_Type/S8.4_A13_T2-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '';'
S8.4_A13_T2

PASS Expected parsing failure

TEST COMPLETE
